Extracted proxy handling to a seperate class. Deja vu
